ingest/requirements.txt
aiobotocore==2.4.2
aiohttp==3.8.3
aioitertools==0.11.0
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
boto3==1.24.59
botocore==1.27.59
charset-normalizer==2.1.1
frozenlist==1.3.3
fsspec==2022.11.0
greenlet==2.0.1
idna==3.4
jmespath==1.0.1
multidict==6.0.4
numpy==1.24.1
pyarrow==10.0.1
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.21.0
s3fs==2022.11.0
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==1.4.45
typing_extensions==4.4.0
urllib3==1.26.13
wrapt==1.14.1
yarl==1.8.2
